requirements.txt
Django==5.0
djangorestframework==3.14.0
django-environ==0.11.2
django-filter==23.5
django-autoslug==1.9.9
django-countries==7.5.1
Pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
mysqlclient==2.2.1
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
gunicorn==21.2.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
Faker==21.0.0
pytest-cov==4.1.0
redis==5.0.1
celery==5.3.6
flower==2.0.1
django-celery-email==3.0.0
argh==0.31.2
PyYAML==6.0.1
watchdog==3.0.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
attrs==23.2.0
drf-spectacular==0.27.1